Registering CosmosPackagesStore in constructor instead of StoreConfig
